Silence warning.

+++ server/TracyView.cpp
@@ -359,7 +359,7 @@ void View::NewZone( Event* zone, uint64_t thread )
     auto it = m_threadMap.find( thread );
     if( it == m_threadMap.end() )
     {
-        m_threadMap.emplace( thread, m_threads.size() );
+        m_threadMap.emplace( thread, (uint32_t)m_threads.size() );
         m_threads.push_back( ThreadData { thread } );
         timeline = &m_threads.back().timeline;
     }
